feat(contract-service): refactor API actions for merchant contract settings and improve error handling
