+++ apps/web/src/actions/unirefund/ContractService/delete-actions.ts
@@ -12,6 +12,17 @@ export async function deleteMerchantContractHeaderByIdApi(id: string) {
   }
 }
 
+export async function deleteMerchantContractContractSettingsByIdApi(id: string) {
+  try {
+    const client = await getContractServiceClient();
+    return structuredResponse(
+      await client.contractsMerchant.deleteApiContractServiceMerchantsContractsContractSettingsById({id}),
+    );
+  } catch (error) {
+    return structuredError(error);
+  }
+}
+
 export async function deleteRefundPointContractHeadersById(id: string) {
   try {
     const client = await getContractServiceClient();
@@ -22,33 +33,3 @@ export async function deleteRefundPointContractHeadersById(id: string) {
     return structuredError(error);
   }
 }
-
-// export async function deleteRefundFeeHeadersById(id: string) {
-//   try {
-//     const client = await getContractServiceClient();
-//     return structuredResponse(
-//       await client.refundFeeHeader.del(
-//         {
-//           id,
-//         },
-//       ),
-//     );
-//   } catch (error) {
-//     return structuredError(error);
-//   }
-// }
-
-// export async function deleteRefundTableHeadersById(id: string) {
-//   try {
-//     const client = await getContractServiceClient();
-//     return structuredResponse(
-//       await client.refundTableHeader.del(
-//         {
-//           id,
-//         },
-//       ),
-//     );
-//   } catch (error) {
-//     return structuredError(error);
-//   }
-// }
